pulled out utils folder to not duplicate code
