+++ src/utils/input.ts
